change points to int type

+++ src/systems/physics_system.h
@@ -5,9 +5,9 @@
 #ifndef PANDAEXPRESS_PHYSICS_SYSTEM_H
 #define PANDAEXPRESS_PHYSICS_SYSTEM_H
 
-static const float BREAD_KILL_POINTS = 50.f;
+static const int BREAD_KILL_POINTS = 50;
 
-static const float LLAMA_KILL_POINTS = 150.f;
+static const int LLAMA_KILL_POINTS = 150;
 
 #include "system.h"
 #include "components/obeys_gravity.h"
